refactor: abstract the quick category filter to a /common component
